Swiss Fertility and Socioeconomic Indicators (1888) Data

Format

A data frame with 47 observations on 6 variables, each of which is in percent, i.e., in 0, 100.

Fertility

Ig, ‘common standardized fertility measure’.

Agriculture

% of males involved in agriculture as occupation.

Examination

% draftees receiving highest mark on army examination.

Education

% education beyond primary school for draftees.

Catholic

% ‘catholic’ (as opposed to ‘protestant’).

Infant.Mortality

live births who live less than 1 year.

@source The data is from the swiss dataset in R.

swiss Standardized fertility measure and socio-economic indicators for each of 47 French-speaking provinces of Switzerland at about 1888. Switzerland, in 1888, was entering a period known as the demographic transition; i.e., its fertility was beginning to fall from the high level typical of underdeveloped countries. The data collected are for 47 French-speaking “provinces” at about 1888. Here, all variables are scaled to 0, 100, where in the original, all but "Catholic" were scaled to 0, 1. datasets